Using prepacked core profile interpolation func
Using the new core profile interpolation function that returns an
inteprolating function instead of values at specific points. This
helped increase the speed of the code such that for relative tolerance
of 10 minutes, most integrations finish withing 10 seconds. However,
the horizontal line of sight chord is taking longer.

The major reason for slow speed of numerical integration is the presence
of peaks that require a lot of fine resolution function calls by quadGK
to reach to the desired relative tolerance. These peaks are naturally
present in the data and we can not do much about it.

There is bit of commented code in this commit where I tried to use multi
processing using threads but that is returning negative inegrated
electron density and is not really speeding up the performance, so I
have commented it out. Maybe in future this can be done more efficiently