Interferometer now functioning fast enough
* Most changes were required in GGDUtils for making interpolations
and projection faster.
* The main change in interferometer.jl was to reuse the calculations
as much as possible.
* I also tried calculating the integrated electron density in each
segment of chord (where each segment is either in edge profile or in
core profile) but this was slower than the current method of simply
integrating over the whole chord and using ∈ to find the relevant
function to use for interpolation.
* I have left the other method as commented out code right now. Will
clean up in next commit to store this work as well.