Fixed bug. Now results are withing 10 ppm.
Fixed the issue that after the optimization fo previous commit, the
integrated electron density was not matching with the previous results.
The value had a fractional error of 0.29289. This was due to a typo in
calculation of dline which was giving answer sqrt(2) times less. After
fixing this bug, the results are withing 10 ppm from previous results.
The difference is probably due to floating point error originating from
GGDUtils, but this calculation might have less error than before.

Memory performance is exactly same as before.
add_interferometer runs with:
```
3.297911 seconds (102.78 k allocations: 532.375 MiB, 2.67% gc time)
```
along with optimization branch on GGDUtils.

Other changes:
- Cleaned up the code a bit
- Made inline overloaded functions for dline and get_line_of_sight
- Added update_chord_core_length function.