Adding check for duplicates in add_interferometer!
This commit resolves https://github.com/ProjectTorreyPines/SynthDiag.jl/issues/17
add_interferometer! function has been modified to read the json file
using JSON.parsefile directly and pass it to its method that takes as
input a dictionary. Note that JSON.parsefile returns a Dict{String, Any}
dictionary while OMAS.imas2dict constructs a Dict{Symbol, Any} dictionary.
Thus after reading the josn file it is converted into a Dict{Symbol, Any}.
Then a duplicacy check is run to ensure that new file has no channels that
have an overlapping name or identifier. If a duplicate is found, an error
is thrown and it recommends the user to use overwrite=true.

In case of overwrite=true, the duplicate channels in existing ids are
deletec and the new channels are added. This may change the indices of
the channel in ids.interferometer.channel array.

In case there is no duplicacy or overwrite=true with additional
channels, the new channels are appended at the end of the
ids.interferometer.channel array.

Test cases have been expanded to cover the new functionality.