Added option to provide latency
Global latency in gas_injection can be set as per defined in IMAS
schema. ids.gas_injection.latency is used for all gas valves unless
they have their own latency defined.
Note that current IMAS schema and IMASDD ids do not have fields for
inidividual gas valve latency. But this can be added in future.
For now, compute_gas_injection! is provided an optional argument
which is a dictionary of valve names and with values as additonal model.
Right now, only latency is read in the additional model. This will be
expaned in future.