Added langmuir_probes_current function
Added langmuir_probes_current function to calculate the current from
the langmuir probes. This also serves as a modular function for doing
calculations. The runtests.jl now plots an IV curve to demonstrate
the use of lamgnuire_probes_current. This function is now used in
the compute_lanngmuir_probes function.

In future, the current calculation can be used for supporting harmonic
method as well.

Useful functions defined in interferometer have been moved to utils.jl
Now, langmuir_probes also support checking of duplication and
overwriting feature.

If v_probe is provided along with v_plasma, a true estimate of measured
langmuir probe current is written in ion_saturation_current field. But
if these are not provided, a bias of -200.0 V is assumed for the probe.