Adding bolometer as synthetic diagnostic. Partial
Bolometer has been partially implemented as a synthetic diagnostic.
This commit adds a significant portion of the code required to do the
bolometer computation. Apart from basic loading of the IDS, the calculation
of the field of view (FoV) of each channel has been implemented. This
calculation returns a conical region defined by a new mutable struct `FoV`.
This is not necessarily the smallest possible FoV, but is a good approximation
to the actual FoV, reducing amount of integration area and time required to
find specific boundaries of FoV. By using an encompassing conical region,
the integral will be a simple computation in cynlindrical coordinates
in the field of view coordiinate frame which can be easily transformed to
the R,Z,Phi coordinate frame.