Fixed errors in transformations and more
The transformations from aperture or detector frames to RPZ frame were
incorrectly implemented. Instead, transformations are now done to
the cartesian XYZ frame and appropriate transformations are defined.

Added functions to add detectors and apertures to channels manually.
This will be sueful for testing as well as any bolometer design
optimization use case.

Also added function to calculate area of the outline and consolidated
duplciate code between aperture and detector using common functions
defined on Union of the two types.