Adding SOS state vector initialization in gas inj
While computing gas injection, when low pass filter is applied to the
flow rate, it would show a transient if the flow rate starts at a
non-zero value. This could be bad for MPC which uses the computation
while optimization. To avoid this, we initialize the state vector of
the low pass filter to the initial value of the flow rate. This way,
there will never be an initial transient due to the low pass filter.