future_flow_rates should be nothing initially
When compute_gas_injection is called with a latency which is longer
than the history of sent commands, then there haven't been enough
commands to calculate future flow rates. In this case, future_flow_rates
should be returned as nothing. This commit mainly brings this feature
so that computer_gas_injection can be called incrementally right
from the start even with a latency.

Also, ensured that if compute_gas_injection is called with a single
value of gas command, it ignore filtering and dribbling as those are
dynamic features and require a command history, and simply return the
flow rate calculated with response curve and respecting latency.