Adding compute_intersection function
New function to compute the intersection of a line in 3D with a
2D surface generated by rotating a line segment in R-Z plane has been
added. This function solves the equations analytically with known
quadratic formula and should be very fast.

Now, each bolometer's FoV, we can determine what section of FoV is
inside the walls of the vacuum vessel. This region will then be
integrated over with the radiaiton emission power density and multiplied
by the illuminated solid angle on the bolometer's detector to get the
total measured power.

Almost there!