Added compute_intersection_s and using SVectors
The computation of intersection of line with conical surface has been
split into a separate function compute_intersection_s. This function
calculates the s value required in line_0 + s * line_dir to intersect
the conical surface. In terms of finding intersection of line of sight
with walls, this will directly correspond to z value in FoV frame and
would be the limits of the intersection.

Other computer intersection functions ahve been modified to return a
single type of StaticVector to make the code faster.